Renamed entities package to song.
