+++ res/view/SongBrowserLayout.fxml
@@ -1,3 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <!-- 
  RapidTunes.
  The music application to help you use all your music sources in one place.
@@ -24,8 +25,6 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  SOFTWARE.
 -->
-
-<?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.*?>
 <?import javafx.scene.lang.*?>
