+++ src/me/oskarmendel/player/search/SearchHandler.java
@@ -32,9 +32,9 @@ import java.util.List;
 
 import com.google.api.services.youtube.model.Video;
 
-import me.oskarmendel.entities.Song;
-import me.oskarmendel.entities.YouTubeSong;
 import me.oskarmendel.player.search.local.LocalSearch;
+import me.oskarmendel.song.Song;
+import me.oskarmendel.song.YouTubeSong;
 
 /**
  * Class that handles all searches. Uses a Singleton.
