+++ src/me/oskarmendel/view/SongBrowserController.java
@@ -42,9 +42,9 @@ import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.paint.Color;
 import javafx.util.Callback;
-import me.oskarmendel.entities.Song;
 import me.oskarmendel.model.CurrentlyPlayingModel;
 import me.oskarmendel.model.SearchResultModel;
+import me.oskarmendel.song.Song;
 
 /**
  * TODO: Create a CellFactory that decides how to present the listView Cells.
